+++ src/cjs/index.cjs
@@ -1,0 +1,10 @@
+/*
+* src/cjs-tools/index.cjs
+*/
+const clearFirestoreData = require('./clearFirestoreData.cjs');
+const prime = require('./prime.cjs');
+
+module.exports = {
+  clearFirestoreData,
+  prime
+}